Rearranged Components folder
Rearranged Components folder, each component has a separate folder. Ran formatter
