+++ client/src/Components/Footer/Footer.jsx
@@ -1,8 +1,8 @@
-import React from 'react'
-import * as Style from './Footer.styles'
+import React from 'react';
+import * as Style from './Footer.styles';
 
 const Footer = () => {
   return <Style.Footer>Weather App &copy; Lukas Vanagas</Style.Footer>;
-}
+};
 
-export default Footer
+export default Footer;
