+++ client/src/Components/Navbar/Navbar.jsx
@@ -1,9 +1,5 @@
-import React from 'react'
-import {
-  Nav,
-  NavLink,
-  ThemeButton
-} from './Navbar.styles'
+import React from 'react';
+import { Nav, NavLink, ThemeButton } from './Navbar.styles';
 const Navbar = () => {
   return (
     <Nav>
@@ -12,7 +8,7 @@ const Navbar = () => {
       <NavLink to="/contact">Contact</NavLink>
       <NavLink to="/cities">Search</NavLink>
     </Nav>
-  )
-}
+  );
+};
 
-export default Navbar
+export default Navbar;
