+++ client/src/Components/SearchBar/Searchbar.styles.js
