+++ client/src/Pages/CitySearch.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import MultiWeather from '../Components/MultiCard';
-import SearchBar from '../Components/SearchBar';
+import MultiWeather from '../Components/MultiCard/MultiCard';
+import SearchBar from '../Components/SearchBar/SearchBar';
 
 const CitySearch = () => {
   return (
